--- packages/buidler-etherscan/.env
@@ -1,1 +1,0 @@
-TS_NODE_FILES=true
