+++ packages/buidler-etherscan/.gitignore
@@ -97,3 +97,4 @@ typings/
 # DynamoDB Local files
 .dynamodb/
 
+.env
